build google test with an external project
Google test is downloaded and built using cmake external project when
USE_SYSTEM_GTEST is set to OFF - the default. It can still be used from the
installed location when setting USE_SYSTEM_GTEST to ON.

The mac homebrew package manager does not provide any way to install google test
and advise to use it on a per project basis.