Cast before and after call to `sqrt`
A call to `sqrt` was taking a `size_t` value as input and expecting a
`size_t` value as output. This kind of works with C++11, but only with
C++11. Even then a `double` is the expected return type. To fix this
issue, we convert to a `double` start with. Also as we should have a
perfect square going into `sqrt`, we expect to have an integral value
coming out (though of `double` type). Then we cast back to `size_t`
explicitly to match our intended storage type.