Fixed static instantiation issue on (at least) Windows.
Previously, the static initialization of FFTRepo::lockRepo in some cases could be uninitialized before the call to clfftSetup was called. This lead to much crashing.

Now, the static function FFTRepo::lockRepo() returns a reference to the local static variable lock. This asserts that the result of lockRepo() must be instantiated in all cases. This also maintains the original intent of lockRepo.